Rename: 프로젝트 파일 구조 재정비
패키지 구조를 정리하고 Controller, Service, DTO, Config 등의 파일 경로를 일관성 있게 재배치함
