--- apac-solution-challenge/apac/src/main/java/com/example/apac/calendar/domain/DailyGoalAchieved.java
@@ -1,14 +1,0 @@
-package com.example.apac.calendar.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class DailyGoalAchieved {
-    private String email;
-    private String date;
-    private int goalAchievedCount;
-}
