+++ apac-solution-challenge/apac/src/main/java/com/example/apac/controller/CalendarController.java
@@ -1,9 +1,9 @@
-package com.example.apac.calendar.controller;
+package com.example.apac.controller;
 
-import com.example.apac.calendar.domain.DailyGoalAchieved;
-import com.example.apac.calendar.dto.CalendarReportResponseDTO;
-import com.example.apac.calendar.dto.CalendarResponseDTO;
-import com.example.apac.calendar.service.CalendarService;
+import com.example.apac.domain.DailyGoalAchieved;
+import com.example.apac.dto.CalendarReportResponseDTO;
+import com.example.apac.dto.CalendarResponseDTO;
+import com.example.apac.service.CalendarService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
@@ -35,10 +35,10 @@ public class CalendarController {
         return calendarService.saveDailyGoal(dailyGoalAchieved);
     }
 
-//    @GetMapping("/report")
-//    public CalendarReportResponseDTO getCalendarReport(Authentication authentication, @RequestParam int year, @RequestParam int month, @RequestParam int day){
-//        String email = authentication.getName();
-//
-//        return calendarService.getCalendarReport(email, year, month, day);
-//    }
+    @GetMapping("/report")
+    public CalendarReportResponseDTO getCalendarReport(Authentication authentication, @RequestParam int year, @RequestParam int month, @RequestParam int day){
+        String email = authentication.getName();
+
+        return calendarService.getCalendarReport(email, year, month, day);
+    }
 }
