+++ apac-solution-challenge/apac/src/test/java/com/example/apac/calendar/calendarControllerTest.java
@@ -1,34 +1,24 @@
 package com.example.apac.calendar;
 
-import com.example.apac.calendar.controller.CalendarController;
-import com.example.apac.calendar.dto.CalendarDayDTO;
-import com.example.apac.calendar.dto.CalendarResponseDTO;
-import com.example.apac.calendar.service.CalendarService;
+import com.example.apac.controller.CalendarController;
+import com.example.apac.dto.CalendarResponseDTO;
+import com.example.apac.dto.HealthDataDto;
+import com.example.apac.service.CalendarService;
 import com.example.apac.security.JwtUtil;
 import com.google.api.Authentication;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.http.HttpHeaders;
-import org.springframework.security.authentication.TestingAuthenticationToken;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 
-import javax.crypto.SecretKey;
-import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.List;
 
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.when;
-import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.authentication;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -56,9 +46,9 @@ public class calendarControllerTest {
         String email = "test@example.com";
         int year = 2025;
         int month = 5;
-        List<CalendarDayDTO> days = Arrays.asList(
-                new CalendarDayDTO("2025-05-01", 2),
-                new CalendarDayDTO("2025-05-02", 3)
+        List<HealthDataDto.CalendarDayDTO> days = Arrays.asList(
+                new HealthDataDto.CalendarDayDTO("2025-05-01", 2),
+                new HealthDataDto.CalendarDayDTO("2025-05-02", 3)
         );
         CalendarResponseDTO responseDTO = new CalendarResponseDTO(year, month, days);
